+++ database/migrations/2024_04_28_071450_claim_request.php
@@ -8,19 +8,23 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('claim_request', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('user_id');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->text('reward_id');
-            $table->text('approved')->nullable();
-            $table->text('pending')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('claim_request')) {
+            Schema::create('claim_request', function (Blueprint $table) {
+                $table->id();
+                $table->uuid('user_id');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->text('reward_id');
+                $table->text('approved')->nullable();
+                $table->text('pending')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::dropIfExists('claim_request');
+        if (Schema::hasTable('claim_request')) {
+            Schema::dropIfExists('claim_request');
+        }
     }
 };
